pyproject.toml
[project]
name = "xxw-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "cryptography>=46.0.5",
    "flask>=3.1.3",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.11.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]
